refactor from page
