+++ src/app/App.tsx
@@ -4,7 +4,7 @@ import MainPage from './components/pages/mainPage/mainPage';
 import AboutUs from './components/pages/aboutUs/aboutUs';
 import ErrorPage from './components/pages/errorPage/errorPage';
 import MainLayout from './layouts/main';
-import CreatedForm from './components/pages/createdForm/createdForm';
+import CreatedFormPage from './components/pages/createdFormPage/createdFormPage';
 class App extends Component {
   render(): ReactNode {
     return (
@@ -13,7 +13,7 @@ class App extends Component {
           <Route path="/" element={<MainLayout />}>
             <Route index element={<MainPage />} />
             <Route path="/about" element={<AboutUs />} />
-            <Route path="/form" element={<CreatedForm />} />
+            <Route path="/form" element={<CreatedFormPage />} />
             <Route path="/404" element={<ErrorPage />} />
             <Route path="*" element={<Navigate to="/404" replace={true} />} />
           </Route>
