+++ src/app/components/pages/createdFormPage/createdForm.module.css
@@ -1,0 +1,6 @@
+.form__title {
+  text-align: center;
+  margin-bottom: 15px;
+  font-size: 24px;
+  font-weight: 500;
+}
